move all of the inner workings to base component
